+++ CHANGELOG.md
@@ -4,6 +4,7 @@ This project changes will be shown here.
 ##[unreleased]
 - Issue 174 - CNC036 - Record backup success as percentage
 - Issue171: CNC035 - Display Calendar link on all activities
+- Issue 172 - CNC036 - Sales Order link not shown when editing an activity
 
 ##[1.2.1]
 - Fix issue with time allocated emails not getting sent